Source code for lesson 3.31

+++ app/Entity/Invoice.php
@@ -27,8 +27,8 @@ class Invoice
     #[Column, GeneratedValue]
     private int $id;
 
-    #[Column(type: Types::DECIMAL, precision: 10, scale: 2)]
-    private float $amount;
+    #[Column(type: Types::BIGINT)]
+    private int $amount;
 
     #[Column(name: 'invoice_number')]
     private string $invoiceNumber;
